+++ packages/form_bloc/lib/src/blocs/form/form_event.dart
@@ -58,3 +58,8 @@ class OnSubmittingFormBloc extends FormBlocEvent {
   @override
   List<Object> get props => [];
 }
+
+class DeleteFormBloc extends FormBlocEvent {
+  @override
+  List<Object> get props => [];
+}