+++ packages/form_bloc/test/text_field/text_field_bloc_test.dart
@@ -136,5 +136,35 @@ void main() {
 
       fieldBloc.clear();
     });
+
+    test('if the initialValue is null, it will be an empty string', () {
+      TextFieldBloc fieldBloc;
+      TextFieldBlocState initialState;
+      List<TextFieldBlocState> expectedStates;
+
+      fieldBloc = TextFieldBloc(initialValue: null);
+
+      initialState = TextFieldBlocState(
+        value: '',
+        error: null,
+        isInitial: true,
+        suggestions: null,
+        isValidated: true,
+        isValidating: false,
+        toStringName: null,
+      );
+
+      expectedStates = [initialState];
+
+      expect(
+        fieldBloc.initialState,
+        initialState,
+      );
+
+      expect(
+        fieldBloc,
+        emitsInOrder(expectedStates),
+      );
+    });
   });
 }